Added sounds, score and different types of blocks.
Added a second level.
A random sound plays when a ball hits any collider.
Each block gives a certain amount of points on destruction. The points are displayed on the screen. The ScoreKeeper class persists through all levels.
There are 4 types of blocks now: one indestructible and three destructible (destroy on 1 hit, 2 hits or 3 hits). The block sprite changes depending on the damage level.